2020-02-17 DB: added option to create factory
