+++ cli/index.php
@@ -6,9 +6,8 @@
  */
 
 // set up class for use
-require 'Constants.php';
-require 'Validate.php';
-use Phpcl\LaminasTools\{Constants,ModuleBuilder,ControllerBuilder,Validate};
+require 'vendor/autoload.php';
+use Phpcl\LaminasTools\{Constants,ModuleBuilder,ControllerBuilder,FactoryBuilder,Validate};
 
 // init vars
 $type       = '';
@@ -17,9 +16,10 @@ $what       = '';
 $baseDir    = '';
 $moduleName = '';
 $controller = '';
+$factory    = '';
 
 if (Validate::checkInputs($argv)) {
-    list($what, $baseDir, $moduleName, $controller) = Validate::getInputs();
+    list($what, $baseDir, $moduleName, $controller, $factory) = Validate::getInputs();
 } else {
     echo Validate::getMessage();
     echo Constants::USAGE;
@@ -27,7 +27,7 @@ if (Validate::checkInputs($argv)) {
 }
 
 // pull in config
-$config = require 'config.php';
+$config = require 'config/config.php';
 
 // detect type
 foreach ($config as $key => $value) {
@@ -46,7 +46,6 @@ if (empty($type)) {
 try  {
     // build module
     if ($what == Constants::BUILD_WHAT[0]) {
-        require 'ModuleBuilder.php';
         $builder = new ModuleBuilder($moduleName, $config[$type]);
         switch ($type) {
             case 'zf3' :
@@ -73,7 +72,6 @@ try  {
         }
     // build controller
     } elseif ($what == Constants::BUILD_WHAT[1]) {
-        require 'ControllerBuilder.php';
         $builder = new ControllerBuilder($moduleName, $config[$type]);
         switch ($type) {
             case 'zf3' :
@@ -89,6 +87,18 @@ try  {
         } else {
             printf(Constants::ERROR_UNABLE, $controller) . "\n";
         }
+    // build factory
+    } elseif ($what == Constants::BUILD_WHAT[2]) {
+        $builder = new FactoryBuilder($moduleName, $config[$type]);
+        switch ($type) {
+            case 'zf3' :
+            case 'lam' :
+                $success = $builder->buildLamMvcFactory($baseDir, $factory);
+                break;
+            default :
+                echo Constants::ERROR_TYPE . "\n";
+        }
+        echo $builder->getOutput();
     }
 } catch (Throwable $t) {
     printf(Constants::ERROR_MSG, get_class($t), $t->getMessage(), $t->getTraceAsString()) . "\n";
